Implemented record queue layout and integrated with dummy data
