+++ src/ui/src/components/serverError/serverError.module.scss
@@ -3,13 +3,13 @@
 .serverError {
   background: $gray-cool-gray-cool-10;
   border-radius: 0.25rem;
-  box-shadow: 0px 4px 8px 0px rgba(0, 0, 0, 0.15);
+  box-shadow: 0 0.25rem 0.5rem 0 rgba(0, 0, 0, 0.15);
   padding: 9.563rem 13.625rem;
 }
 
 .instructions {
   border-radius: 0.25rem;
   font-size: 1.125rem;
-  border: solid 1px $red-70;
+  border: solid 0.063rem $red-70;
   background: $gray-cool-gray-cool-20;
 }
