+++ src/ui/src/styles/uswds-customized.scss
@@ -80,59 +80,59 @@ h5 {
 }
 
 .usa-table--record-linker {
-  border-radius: 4px;
-  box-shadow: 0px 4px 8px 0px rgba(0, 0, 0, 0.15);
+  border-radius: 0.25rem;
+  box-shadow: 0 0.25rem 0.5rem 0 rgba(0, 0, 0, 0.15);
   border-collapse: separate !important;
   background-color: transparent;
 
   th {
-    border: solid 1px #a9aeb1;
-    border-left: 0px;
-    border-right: 0px;
+    border: solid 0.063rem #a9aeb1;
+    border-left: 0;
+    border-right: 0;
     background-color: #dfe1e2;
-    padding: 10px 24px;
+    padding: 0.625rem 1.5rem;
     text-transform: uppercase;
 
     &:first-child {
-      border-top-left-radius: 4px;
-      border-left: solid 1px #a9aeb1;
-      padding-left: 34px;
+      border-top-left-radius: 0.25rem;
+      border-left: solid 0.063rem #a9aeb1;
+      padding-left: 2.125rem;
     }
 
     &:last-child {
-      border-top-right-radius: 4px;
-      border-right: solid 1px #a9aeb1;
-      padding-right: 34px;
+      border-top-right-radius: 0.25rem;
+      border-right: solid 0.063rem #a9aeb1;
+      padding-right: 2.125rem;
     }
   }
 
   td {
     background-color: #fff;
-    padding: 19px 24px;
-    border: solid 1px #a9aeb1;
+    padding: 1.188rem 1.5rem;
+    border: solid 0.063rem #a9aeb1;
     border-left: 0;
     border-right: 0;
     border-top: 0;
 
     &:first-child {
-      border-left: solid 1px #a9aeb1;
-      padding-left: 34px;
+      border-left: solid 0.063rem #a9aeb1;
+      padding-left: 2.125rem;
     }
 
     &:last-child {
-      border-right: solid 1px #a9aeb1;
-      padding-right: 34px;
+      border-right: solid 0.063rem #a9aeb1;
+      padding-right: 2.125rem;
     }
   }
 
   tr:last-child {
     td {
       &:first-child {
-        border-bottom-left-radius: 4px;
+        border-bottom-left-radius: 0.25rem;
       }
 
       &:last-child {
-        border-bottom-right-radius: 4px;
+        border-bottom-right-radius: 0.25rem;
       }
     }
   }
